+++ AoC_2023/IDaySolver.cs
@@ -2,6 +2,6 @@ namespace AoC_2023;
 
 internal interface IDaySolver
 {
-    Task<int> SolveAsync();
-    Task<int> SolveBonusAsync();
+    Task<long> SolveAsync();
+    Task<long> SolveBonusAsync();
 }
